+++ packages/shared/src/services/keeperService.ts
@@ -1,6 +1,6 @@
+import type { ZkProofResponse } from '@evevault/shared/types/zkLogin'
 import { Ed25519PublicKey } from '@mysten/sui/keypairs/ed25519'
 import type { SuiChain } from '@mysten/wallet-standard'
-import type { ZkProofResponse } from '#/types/enoki'
 import { VaultMessageTypes, type VaultResponse } from '#/types/messages'
 import type { StoredSecretKey } from '#/types/stores'
 import { VAULT_UNLOCK_MS } from '#/utils/constants'
