+++ packages/shared/src/services/vaultService.ts
@@ -1,7 +1,7 @@
 import type { ZKWebCryptoSigner } from '@evefrontier/wallet-core/crypto'
+import type { ZkProofResponse } from '@evevault/shared/types/zkLogin'
 import type { PublicKey } from '@mysten/sui/cryptography'
 import type { SuiChain } from '@mysten/wallet-standard'
-import type { ZkProofResponse } from '#/types/enoki'
 import type { StoredSecretKey } from '#/types/stores'
 import { createWebCryptoPlaceholder } from '#/types/wallet'
 import { isWeb } from '#/utils/environment'
