+++ packages/shared/src/types/index.ts
@@ -1,9 +1,9 @@
 export * from './actions'
 export * from './authTypes'
 export * from './components'
-export * from './enoki'
 export * from './messages'
 export * from './networks'
 export * from './stores'
 export * from './tenant'
 export * from './wallet'
+export * from './zkLogin'
